Add files via upload
This program is a Python application built using the tkinter library for creating a graphical user interface (GUI). It provides buttons that allow users to navigate to the account deletion or deactivation pages of various social media platforms, such as X, Instagram, Facebook, and more. The GUI includes a language change feature that toggles between English and Spanish, dynamically updating the button labels accordingly. Additionally, the program incorporates sound effects using the pygame library, playing a sound when the mouse hovers over a button. Overall, the program provides a user-friendly interface for managing social media account deletions with interactive elements and language flexibility.